requirements.txt
# core
django==4.0.4
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0


# Development
pytest>=7.1.1,<7.2.0
mypy==0.942
black==22.3.0
